Align HdrHistogram to C version
Refactor hdr package, so that it is more closely aligned to
HdrHistogram_c. The advantage being that upstream bug fixes are easier
to track and integrate into the C++ port.

Statistical functions that depend on iterators have been moved to a
separate file in order to break a cyclic dependency between files.

Improve unit-test coverage has also been improved in this commit.