feat(log): support for asynchronous loggers
Update logger components so that the underlying storage that backs the
log stream buffer is dynamically allocated.

The dynamically allocated memory block can then be detached from the
stream buffer and ownership transferred to the logger implementation.

The benefit of this approach is that it allows for efficient logger
implementations that process the log message asynchronously on a
background thread.

DEV-1325