perf(fix): Use boost vector as buffer instead of std::vector
Buffer::prepare() is called every time when sth is about to be written
to the buffer, it includes calling of vector::resize().

It turned out that resize() is not so cheap if current capacity is
more than requested size, because it touches newly requested chunk of
memory.
Measurements showed that boost::vector does it a bit more efficient
by using memset() while std::vector uses std::__uninitialized_default_n_a()

SDB-6760