fix(util): improve dec_digits benchmarks and move pow10
Previously, dec_digits was invoked with inputs starting from 0 and
increasing by 1. This is a very predictable pattern, and easy for the
branch predictor to learn. Thus, the benchmarks show unrepresentative
performance and favour branchy implementations of dec_digits.

Now, dec_digits is invoked with random numbers of mixed digits.
pow10 was previously in reactive-cpp but moved as it's now required in
toolbox