feat(timer): Allow limiting the work done by dispatch
Sometimes it is required that TimerQueue::dispatch finishes fast in a
bounded amount of time while still making progress on the queued timers.
In which case, you want dispatch to process a small number of queued
timers and then return back to caller.

Furthermore, limit the work done by low priority timer queue to just 1
in the reactor. This prevents too much time being taken by low priority
timers if there are a lot of them eligible to be dispatched.
They are low priority timers, so its ok if they are delayed by a little
bit of time.

SDB-8174