chore: Remove catch-all overload of OStaticStream
This an issue because if we attempt to create an overloaded operator<<
for a custom type then it results in ambiguous call compiler error as
there are now 2 valid overloads (first is catch-all in toolbox, and 2nd
our custom defined one).

The catch-all overload mainly exists to ensure that after several
streaming operations, OStaticStream& (ref) is returned. This then allows
the OStaticStream& to be implicity converted to a string. For example,
consider an expression like so: err_msg() << a << b << c;

If the catch-all overload did not exist, then this expression would
return a 'std::ostream&' rather than 'OStaticStream&'. This would
prevent the stream from being turned into a string implicitly as
std::ostream& is not implicitly convertible to a string_view.

It is only 'err_msg()' that relies on this behaviour. Thus, 'ErrMsg'
has been promoted to a strong type, and implements the functionality
it requires rather than burdening a generic type like OStaticStream.
And so, the catch-all overload of OStaticStream, as well as implicit
conversion to string_view can be removed.

SDB-8506