chore: Replace use of ostream IO manipulators
Currently, our streaming overloads (operator<<) make use of io
manipulators. This means any ostreams making use of such overloads
are also required to implement IO manipulators -- this adds non-trivial
complexity and state. Not only that, they are cumbersome to use, and
typically require saving existing state and restoring it with an RAII
object -- this is often forgotten, as is the case in our codebase.

The replacment is built on top of C++20's std::format.

SDB-8506