feat: Allow custom high priority reactor hook
The Reactor processes pending events in a "reactor cycle". A reactor
cycle may have a lot of events to process, for example during market
data bursts. Each of these events may take an arbitrarily long time to
be processed. Currently, while these events are being processed, they
will periodically "yield" back control to the Reactor. This gives the
Reactor a chance to poll for high priority events. Frequently polling
for high-priority events reduce latency outliers as it ensures that high
priority events are not left pending for too long. This works great if
the program has only one "source of events" (i.e. the Reactor).

Some processes like LPFIX/CLFIXGW have multiple "source of events",
i.e. IPC agent + Reactor. These processes need to poll both of these
"source of events" frequently. e.g. LPFIX receives events via IPC (from
Crossfire) + via the Reactor (from LPFIXGW).

The issue with the Reactor is that it can take arbitrarily long when
it's polled. For example, during market data bursts, polling the reactor
can take a long time if there are lots of messages to process.

When this happens the IPC agent can be starved as it hasn't been polled
for a long time, and importantly the IPC agent may have higher-priority
events to process than the Reactor (e.g. orders).

The existing "yielding" mechanism is extended -- allowing a custom user
hook to be installed. The intention is to use this hook to poll the IPC
agent.
So, when the Reactor normally polls for high priority events, it will
also invoke the user hook so that any user defined high priority actions
can be carried out.

SDB-9450